+++ app_plus/src/main/java/org/gkisalatiga/plus/lib/AppPreferences.kt
@@ -24,9 +24,6 @@ class AppPreferences(private val ctx: Context) {
     private val DEFAULT_PREFERENCE_KEY_VALUES: Map<PreferenceKeys, Any> = mapOf(
         PreferenceKeys.PREF_KEY_CAROUSEL_BANNER_UPDATE_FREQUENCY to 86400000.toLong(),  // --- 86400000 means "once every 1 day" in millisecond
         PreferenceKeys.PREF_KEY_KEEP_NUMBER_OF_CACHED_PDF_FILES to 30.toInt(),
-        PreferenceKeys.PREF_KEY_LAST_STATIC_DATA_UPDATE to Long.MIN_VALUE,
-        PreferenceKeys.PREF_KEY_LAST_CAROUSEL_BANNER_UPDATE to Long.MIN_VALUE,
-        PreferenceKeys.PREF_KEY_LAUNCH_COUNTS to -1.toInt(),
         PreferenceKeys.PREF_KEY_STATIC_DATA_UPDATE_FREQUENCY to 604800000.toLong(),  // --- 604800000 means "once every 7 days" in millisecond
     )
 
@@ -142,8 +139,5 @@ class AppPreferencesCompanion : Application () {
 enum class PreferenceKeys {
     PREF_KEY_CAROUSEL_BANNER_UPDATE_FREQUENCY,
     PREF_KEY_KEEP_NUMBER_OF_CACHED_PDF_FILES,
-    PREF_KEY_LAST_STATIC_DATA_UPDATE,
-    PREF_KEY_LAST_CAROUSEL_BANNER_UPDATE,
-    PREF_KEY_LAUNCH_COUNTS,
     PREF_KEY_STATIC_DATA_UPDATE_FREQUENCY,
 }